third_party/blink/web_tests/external/wpt/html/editing/dnd/interactiveelements/input_image_inside_draggable_element-expected.txt

This is a testharness.js-based test.
[FAIL] Input image button inside draggable element should be draggable from border
  assert_equals: should drag outer element expected "outer" but got "inner"
[FAIL] Input image button inside draggable element should be draggable from image
  assert_equals: should drag outer element expected "outer" but got "inner"
Harness: the test ran to completion.
